增加登出功能，调整用户界面

+++ OrzBili/ApiModels/BangumiDetailModel/BangumiDetialModel.cs
@@ -92,7 +92,7 @@ public class Progress
 
 public class Vip_Info
 {
-    public int? due_date { get; set; } = default;
+    public long? due_date { get; set; } = default;
 
     public int? status { get; set; } = default;
